refactor(rust): reuse the abac control policy inside the policy access control
