+++ examples/rust/get_started/examples/11-attribute-based-authentication-control-plane.rs
@@ -1,12 +1,13 @@
 use hello_ockam::{create_token, get_credential, import_project};
-use ockam::access_control::{AbacAccessControl, AllowAll};
 use ockam::identity::authenticated_storage::AuthenticatedAttributeStorage;
 use ockam::identity::credential::{Credential, OneTimeCode};
 use ockam::identity::{Identity, TrustEveryonePolicy, TrustMultiIdentifiersPolicy};
 
+use ockam::abac::AbacAccessControl;
 use ockam::remote::RemoteForwarder;
 use ockam::{route, vault::Vault, Context, Result, TcpTransport};
 use ockam_api::DefaultAddress;
+use ockam_core::AllowAll;
 use std::sync::Arc;
 use std::time::Duration;
