--- implementations/rust/ockam/ockam/src/access_control/mod.rs
@@ -1,9 +1,0 @@
-//! Access Control re-exports and implementations
-//!
-//! This is where we can find access control implementations using
-//! several dependencies of the ockam crate like the ockam_abac and ockam_identity crates
-mod attribute_access_control;
-
-pub use attribute_access_control::*;
-pub use ockam_core::access_control::*;
-pub use ockam_identity::access_control::*;
