+++ implementations/rust/ockam/ockam_abac/src/lib.rs
@@ -13,9 +13,11 @@ mod types;
 #[cfg(feature = "std")]
 mod parser;
 
+pub mod attribute_access_control;
 pub mod expr;
 pub mod mem;
 
+pub use attribute_access_control::AbacAccessControl;
 pub use env::Env;
 pub use error::{EvalError, ParseError};
 pub use eval::eval;
