+++ implementations/rust/ockam/ockam_abac/src/policy.rs
@@ -1,18 +1,13 @@
-use core::{fmt, str};
+use core::fmt;
 use ockam_core::compat::boxed::Box;
-use ockam_core::compat::format;
-use ockam_core::compat::string::ToString;
 use ockam_core::{async_trait, RelayMessage};
 use ockam_core::{IncomingAccessControl, Result};
-use ockam_identity::{
-    authenticated_storage::IdentityAttributeStorage, IdentitySecureChannelLocalInfo,
-};
+use ockam_identity::authenticated_storage::IdentityAttributeStorage;
 use tracing as log;
 
-use crate::eval::eval;
-use crate::expr::str;
 use crate::traits::PolicyStorage;
 use crate::types::{Action, Resource};
+use crate::AbacAccessControl;
 use crate::{Env, Expr};
 
 /// Evaluates a policy expression against an environment of attributes.
@@ -26,7 +21,6 @@ pub struct PolicyAccessControl<P, S> {
     policies: P,
     attributes: S,
     environment: Env,
-    overwrite: bool,
 }
 
 impl<P, S> PolicyAccessControl<P, S> {
@@ -42,12 +36,7 @@ impl<P, S> PolicyAccessControl<P, S> {
             policies,
             attributes: store,
             environment: env,
-            overwrite: false,
         }
-    }
-
-    pub fn overwrite(&mut self) {
-        self.overwrite = true
     }
 }
 
@@ -81,95 +70,12 @@ where
             return Ok(false);
         };
 
-        // Get identity identifier from message metadata:
-        let id = if let Ok(info) = IdentitySecureChannelLocalInfo::find_info(msg.local_message()) {
-            info.their_identity_id().clone()
-        } else {
-            log::debug! {
-                resource = %self.resource,
-                action   = %self.action,
-                "identity identifier not found; access denied"
-            }
-            return Ok(false);
-        };
-
-        let mut environment = self.environment.clone();
-
-        // Get identity attributes and populate the environment:
-        if let Some(attrs) = self.attributes.get_attributes(&id).await? {
-            for (key, value) in attrs.attrs() {
-                if key.find(|c: char| c.is_whitespace()).is_some() {
-                    log::warn! {
-                        resource = %self.resource,
-                        action   = %self.action,
-                        id       = %id,
-                        key      = %key,
-                        "attribute key with whitespace ignored"
-                    }
-                }
-                match str::from_utf8(value) {
-                    Ok(s) => {
-                        if !self.overwrite && environment.contains(key) {
-                            log::debug! {
-                                resource = %self.resource,
-                                action   = %self.action,
-                                id       = %id,
-                                key      = %key,
-                                "attribute already present"
-                            }
-                            continue;
-                        }
-                        environment.put(format!("subject.{key}"), str(s.to_string()));
-                    }
-                    Err(e) => {
-                        log::warn! {
-                            resource = %self.resource,
-                            action   = %self.action,
-                            id       = %id,
-                            key      = %key,
-                            err      = %e,
-                            "failed to interpret attribute as string"
-                        }
-                    }
-                }
-            }
-        };
-
-        //add the identifier itself as a subject parameter
-        environment.put("subject.identifier", str(id.to_string()));
-
-        // Finally, evaluate the expression and return the result:
-        match eval(&expr, &environment) {
-            Ok(Expr::Bool(b)) => {
-                log::debug! {
-                    resource      = %self.resource,
-                    action        = %self.action,
-                    id            = %id,
-                    is_authorized = %b,
-                    "policy evaluated"
-                }
-                Ok(b)
-            }
-            Ok(x) => {
-                log::warn! {
-                    resource = %self.resource,
-                    action   = %self.action,
-                    id       = %id,
-                    expr     = %x,
-                    "evaluation did not yield a boolean result"
-                }
-                Ok(false)
-            }
-            Err(e) => {
-                log::warn! {
-                    resource = %self.resource,
-                    action   = %self.action,
-                    id       = %id,
-                    err      = %e,
-                    "policy evaluation failed"
-                }
-                Ok(false)
-            }
-        }
+        AbacAccessControl::new(
+            self.attributes.async_try_clone().await?,
+            expr,
+            self.environment.clone(),
+        )
+        .is_authorized(msg)
+        .await
     }
 }
